# docs/swagger.yaml
basePath: /
definitions:
  domain.AddressInputDTO:
    properties:
      city:
        type: string
      complement:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - neighborhood
    - number
    - state
    - street
    type: object
...
  domain.CreateProposalRequestDTO:
    properties:
      category_id:
        type: integer
      description:
        type: string
      helper_id:
        type: string
      value:
        type: number
    required:
    - category_id
    - description
    - helper_id
    - value
    type: object
...
  domain.LoginRequestDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResponseDTO:
    properties:
      token:
        type: string
    type: object
...
  domain.ProposalResponseDTO:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      helper_id:
        type: string
      id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      value:
        type: number
    type: object
  domain.RegisterRequestDTO:
    properties:
      address:
        $ref: '#/definitions/domain.AddressInputDTO'
      categories:
        items:
          type: integer
        minItems: 1
        type: array
      document:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      user_type:
        type: string
    required:
    - address
    - categories
    - document
    - email
    - name
    - password
    - phone
    - user_type
    type: object
  domain.RegisterResponseDTO:
    properties:
      categories:
        items:
          type: integer
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      user_type:
        type: string
    type: object
...
  domain.UpdateProposalStatusRequestDTO:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Helptrix helper marketplace platform.
  title: Helptrix API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password, returns a Paseto token
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a helper or business user with address and categories
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.RegisterResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /proposal:
    get:
      consumes:
      - application/json
      description: Lists proposals for the authenticated user. Business users see
        proposals they created; helpers see proposals directed at them. Supports optional
        status filter.
      parameters:
      - description: Filter by proposal status (pending, accepted, refused, in progress,
          cancelled, finished)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ProposalResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List proposals
      tags:
      - proposal
    post:
      consumes:
      - application/json
      description: Creates a new service proposal from a business user directed at
        a helper. Only business users can create proposals.
      parameters:
      - description: Create proposal request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProposalRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ProposalResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a proposal
      tags:
      - proposal
  /proposal/{id}:
    get:
      consumes:
      - application/json
      description: Returns a proposal by ID. Only participants (business owner or
        helper) can access it.
      parameters:
      - description: Proposal ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProposalResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get proposal by ID
      tags:
      - proposal
  /proposal/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of a proposal following the allowed state machine
        transitions. Helpers manage most transitions; both parties can cancel.
      parameters:
      - description: Proposal ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update status request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProposalStatusRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProposalResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update proposal status
      tags:
      - proposal
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"